RT-VIBE/requirements.txt
tqdm
yacs==0.1.6
# h5py==2.10.0
# numpy==1.17.5
# scipy==1.4.1
# numba==0.47.0
smplx==0.1.26
gdown
PyYAML==5.3.1
joblib==0.14.1
# pillow==7.1.0
trimesh==3.5.25
pyrender
progress==1.5
filterpy==1.4.5
# matplotlib==3.1.3
# tensorflow==1.15.4
# tensorboard==1.15.0
# torch==1.11.0
pytube==12.1.0
torchvision==0.12.0
# scikit-image==0.16.2
# scikit-video==1.1.11
# opencv-python==4.1.2.30
# llvmlite==0.32.1
pyzmq
git+https://github.com/mattloper/chumpy.git
git+https://github.com/zc402/yolov3-pytorch.git
git+https://github.com/zc402/multi-person-tracker.git
